feat(Searches): Allow for system search and distance calculations

+++ src/main/groovy/org/jorgetrujillo/elitebot/domain/elite/System.groovy
@@ -20,9 +20,9 @@ class System {
   Boolean needsPermit
 
   // Location info
-  Long x
-  Long y
-  Long z
+  Double x
+  Double y
+  Double z
 
   // Stations
   List<Station> stations
